pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "logistech-openenv"
version = "1.0.0"
description = "LogisTech OpenEnv – Autonomous Global Supply Chain Environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.29.0",
    "pydantic==2.7.1",
    "openai==1.30.1",
    "python-dotenv==1.0.1",
    "pydantic-settings==2.2.1",
    "pymongo==4.7.2",
    "motor==3.4.0",
    "dnspython==2.6.1",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "pyjwt==2.8.0",
    "requests==2.31.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
